+++ src/Saunter/AsyncApiEndpointRouteBuilderExtensions.cs
@@ -21,7 +21,7 @@ namespace Saunter
                 .UseMiddleware<AsyncApiMiddleware>()
                 .Build();
 
-            var options = endpoints.ServiceProvider.GetService<IOptions<AsyncApiOptions>>();
+            var options = endpoints.ServiceProvider.GetRequiredService<IOptions<AsyncApiOptions>>();
             var route = options.Value.Middleware.Route;
 
             return endpoints.MapGet(route, pipeline);
@@ -44,7 +44,7 @@ namespace Saunter
                 .UseMiddleware<AsyncApiUiMiddleware>()
                 .Build();
 
-            var options = endpoints.ServiceProvider.GetService<IOptions<AsyncApiOptions>>();
+            var options = endpoints.ServiceProvider.GetRequiredService<IOptions<AsyncApiOptions>>();
             var route = options.Value.Middleware.UiBaseRoute + "{*wildcard}";
 
             return endpoints.MapGet(route, pipeline);