+++ src/Saunter/AsyncApiServiceCollectionExtensions.cs
@@ -17,15 +17,19 @@ namespace Saunter
         /// <param name="services">The collection to add services to.</param>
         /// <param name="setupAction">An action used to configure the AsyncAPI options.</param>
         /// <returns>The service collection so additional calls can b e chained.</returns>
-        public static IServiceCollection AddAsyncApiSchemaGeneration(this IServiceCollection services, Action<AsyncApiOptions> setupAction = null)
+        public static IServiceCollection AddAsyncApiSchemaGeneration(this IServiceCollection services, Action<AsyncApiOptions>? setupAction = null)
         {
             services.AddOptions();
 
             services.TryAddSingleton<IAsyncApiDocumentCloner, AsyncApiDocumentSerializeCloner>();
             services.TryAddSingleton<IAsyncApiSchemaGenerator, AsyncApiSchemaGenerator>();
-            services.TryAddSingleton<IAsyncApiDocumentProvider, AttributeDocumentProvider>();
 
-            if (setupAction != null) services.Configure(setupAction);
+            services.TryAddTransient<IAsyncApiDocumentProvider, AttributeDocumentProvider>();
+
+            if (setupAction != null)
+            {
+                services.Configure(setupAction);
+            }
 
             return services;
         }